Upgrading to SP15
Merging in the search bar into the SP15 version of the innvoator.aspx file which only contained slight differences. I also changed the CSS to remove the border around the textbox and instead place a border around the parent DIV. This was done to make the look of the search bar more consistent across browsers.